[V1][TPU] Add a TPU-optimized top-p algorithm.
Top-k and top-p are slow on TPU because existing algorithms use torch.scatter.
For some reason torch.scatter is extremely slow on TPU. There's ongoing work
to optimize it, but until that's done, we need an alternative algorithm that
circumvents scattering.

The algorithm in this PR avoids using torch.scatter by finding a "cut-off"
element in the original logit, and after thresholding the logit using this
cut-off, the remaining elements shall constitute the top-p set.

A caveat of the above approach is that ties are not correctly handled --
if there are duplicate cutoff elements present in the logit, then the
resulting top-p set will be incorrect. To address this problem, we
introduce a tiny perturbation to the probabilities (after softmax) to
break any potential ties. The added perturbation is tiny so it should
not alter the end results significantly, but it still makes this algorithm
approximate rather than an exact one.

Signed-off-by: Hyesoo Yang <hyeygit@gmail.com>